+++ pcs/test/test_constraints.py
@@ -22,6 +22,10 @@ from pcs.test.tools.misc import (
 )
 from pcs.test.tools.pcs_runner import pcs, PcsRunner
 
+from pcs.constraint import LOCATION_NODE_VALIDATION_SKIP_MSG
+
+LOCATION_NODE_VALIDATION_SKIP_WARNING = f"Warning: {LOCATION_NODE_VALIDATION_SKIP_MSG}\n"
+
 CONSTRAINTS_TMP = rc("test_constraints")
 if not os.path.exists(CONSTRAINTS_TMP):
     os.makedirs(CONSTRAINTS_TMP)
@@ -295,7 +299,7 @@ Ticket Constraints:
     def testAllConstraints(self):
         self.fixture_resources()
         output, returnVal = pcs(temp_cib, "constraint location D5 prefers node1")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint order Master then D5")
         assert returnVal == 0 and output == "Adding Master D5 (kind: Mandatory) (Options: first-action=start then-action=start)\n", output
@@ -315,16 +319,16 @@ Ticket Constraints:
     def testLocationConstraints(self):
         self.fixture_resources()
         output, returnVal = pcs(temp_cib, "constraint location D5 prefers node1")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint location D5 avoids node1")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint location D5 prefers node1")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint location D5 avoids node2")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint")
         assert returnVal == 0
@@ -337,10 +341,10 @@ Ticket Constraints:
     def testConstraintRemoval(self):
         self.fixture_resources()
         output, returnVal = pcs(temp_cib, "constraint location D5 prefers node1")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint location D6 prefers node1")
-        assert returnVal == 0 and output == "", output
+        assert returnVal == 0 and output == LOCATION_NODE_VALIDATION_SKIP_WARNING, output
 
         output, returnVal = pcs(temp_cib, "constraint remove blahblah")
         assert returnVal == 1 and output.startswith("Error: Unable to find constraint - 'blahblah'"), output
@@ -579,11 +583,11 @@ Colocation Constraints:
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint location add my_constraint_id crd my_node -INFINITY resource-discovery=always")
-        ac(o,"Cluster CIB has been upgraded to latest version\n")
+        ac(o, "Cluster CIB has been upgraded to latest version\n" + LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint location add my_constraint_id2 crd1 my_node -INFINITY resource-discovery=never")
-        ac(o,"")
+        ac(o, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint --full")
@@ -818,10 +822,10 @@ Ticket Constraints:
     def testLocationConstraintRule(self):
         self.fixture_resources()
         o, r = pcs(temp_cib, "constraint location D1 prefers rh7-1")
-        assert r == 0 and o == "", o
+        assert r == 0 and o == LOCATION_NODE_VALIDATION_SKIP_WARNING, o
 
         o, r = pcs(temp_cib, "constraint location D2 prefers rh7-2")
-        assert r == 0 and o == "", o
+        assert r == 0 and o == LOCATION_NODE_VALIDATION_SKIP_WARNING, o
 
         o, r = pcs(temp_cib, "constraint rule add location-D1-rh7-1-INFINITY #uname eq rh7-1")
         assert r == 0 and o == "", o
@@ -1046,7 +1050,7 @@ Warning: changing a monitor operation interval from 10s to 11 to make the operat
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint location stateful1 prefers rh7-1 --force")
-        ac(o,"")
+        ac(o, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint location statefulG rule #uname eq rh7-1 --force")
@@ -1115,13 +1119,13 @@ Warning: changing a monitor operation interval from 10s to 11 to make the operat
         output, returnVal = pcs(temp_cib,
             "constraint location stateful1 prefers rh7-1 --autocorrect"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib,
             "constraint location statefulG prefers rh7-1 --autocorrect"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib,
@@ -1410,7 +1414,7 @@ Ticket Constraints:
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint location dummy prefers rh7-1 --force")
-        ac(o,"")
+        ac(o, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert r == 0
 
         o,r = pcs(temp_cib, "constraint location dummyG rule #uname eq rh7-1 --force")
@@ -1478,13 +1482,13 @@ Ticket Constraints:
         output, returnVal = pcs(temp_cib,
             "constraint location dummy prefers rh7-1 --autocorrect"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib,
             "constraint location dummyG prefers rh7-1 --autocorrect"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib,
@@ -1718,7 +1722,7 @@ Ticket Constraints:
         shutil.copy(large_cib, temp_cib)
 
         output, returnVal = pcs(temp_cib, "constraint location dummy prefers rh7-1")
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert returnVal == 0
 
         output, returnVal = pcs(temp_cib, "constraint location show resources dummy --full")
@@ -1746,7 +1750,7 @@ Ticket Constraints:
         assert returnVal == 0
 
         output, returnVal = pcs(temp_cib, "constraint location dummy prefers rh7-1")
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert returnVal == 0
 
         output, returnVal = pcs(temp_cib, "constraint location show resources dummy --full")
@@ -1760,7 +1764,7 @@ Ticket Constraints:
     def testConstraintResourceCloneUpdate(self):
         self.fixture_resources()
         output, returnVal = pcs(temp_cib, "constraint location D1 prefers rh7-1")
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert returnVal == 0
 
         output, returnVal = pcs(temp_cib, "constraint colocation add D1 with D5")
@@ -1805,7 +1809,7 @@ Ticket Constraints:
         assert returnVal == 0
 
         output, returnVal = pcs(temp_cib, "constraint location DG prefers rh7-1")
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         assert returnVal == 0
 
         output, returnVal = pcs(temp_cib, "constraint colocation add DG with D5")
@@ -1863,25 +1867,25 @@ Ticket Constraints:
         output, returnVal = pcs(
             temp_cib, "constraint location D1 prefers node1=100"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(
             temp_cib, "constraint location D1 prefers guest1=200"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(
             temp_cib, "constraint location D2 avoids node2=300"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(
             temp_cib, "constraint location D2 avoids guest1=400"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib, "constraint --full")
@@ -1940,13 +1944,13 @@ Ticket Constraints:
         output, returnVal = pcs(
             temp_cib, "constraint location D1 prefers guest1=200"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(
             temp_cib, "constraint location D2 avoids guest1=400"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib, "constraint --full")
@@ -2014,7 +2018,7 @@ Ticket Constraints:
         output, returnVal = pcs(
             temp_cib, "constraint location vm-guest1 prefers node1"
         )
-        ac(output, "")
+        ac(output, LOCATION_NODE_VALIDATION_SKIP_WARNING)
         self.assertEqual(0, returnVal)
 
         output, returnVal = pcs(temp_cib, "resource delete vm-guest1")
@@ -2789,7 +2793,8 @@ class LocationTypeId(ConstraintEffect):
                 <rsc_location id="location-A-node1-INFINITY" node="node1"
                     rsc="A" score="INFINITY"
                 />
-            </constraints>"""
+            </constraints>""",
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_avoids(self):
@@ -2804,7 +2809,8 @@ class LocationTypeId(ConstraintEffect):
                 <rsc_location id="location-A-node1--INFINITY" node="node1"
                     rsc="A" score="-INFINITY"
                 />
-            </constraints>"""
+            </constraints>""",
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_add(self):
@@ -2817,7 +2823,8 @@ class LocationTypeId(ConstraintEffect):
             ],
             """<constraints>
                 <rsc_location id="my-id" node="node1" rsc="A" score="INFINITY"/>
-            </constraints>"""
+            </constraints>""",
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_rule(self):
@@ -2847,11 +2854,13 @@ class LocationTypePattern(ConstraintEffect):
     # complete test of location constraints. Instead it relies on legacy tests
     # to test location constraints with plain resource name.
     empty_cib = rc("cib-empty-2.6.xml")
+    _stdout = ""
 
     def stdout(self):
-        return ""
+        return self._stdout
 
     def test_prefers(self):
+        self._stdout += LOCATION_NODE_VALIDATION_SKIP_WARNING
         self.assert_effect(
             "constraint location regexp%res_[0-9] prefers node1",
             """<constraints>
@@ -2863,6 +2872,7 @@ class LocationTypePattern(ConstraintEffect):
         )
 
     def test_avoids(self):
+        self._stdout += LOCATION_NODE_VALIDATION_SKIP_WARNING
         self.assert_effect(
             "constraint location regexp%res_[0-9] avoids node1",
             """<constraints>
@@ -2874,6 +2884,7 @@ class LocationTypePattern(ConstraintEffect):
         )
 
     def test_add(self):
+        self._stdout += LOCATION_NODE_VALIDATION_SKIP_WARNING
         self.assert_effect(
             "constraint location add my-id regexp%res_[0-9] node1 INFINITY",
             """<constraints>
@@ -2903,9 +2914,7 @@ class LocationTypePattern(ConstraintEffect):
 @skip_unless_location_rsc_pattern
 class LocationTypePatternWithCibUpgrade(LocationTypePattern):
     empty_cib = rc("cib-empty-2.0.xml")
-
-    def stdout(self):
-        return "Cluster CIB has been upgraded to latest version\n"
+    _stdout = "Cluster CIB has been upgraded to latest version\n"
 
 
 @skip_unless_location_rsc_pattern
@@ -3183,7 +3192,8 @@ class BundleLocation(Bundle):
                         rsc="B" score="INFINITY"
                     />
                 </constraints>
-            """
+            """,
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_bundle_avoids(self):
@@ -3195,7 +3205,8 @@ class BundleLocation(Bundle):
                         rsc="B" score="-INFINITY"
                     />
                 </constraints>
-            """
+            """,
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_bundle_location(self):
@@ -3205,7 +3216,8 @@ class BundleLocation(Bundle):
                 <constraints>
                     <rsc_location id="id" node="node1" rsc="B" score="100" />
                 </constraints>
-            """
+            """,
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_primitive_prefers(self):
@@ -3226,7 +3238,8 @@ class BundleLocation(Bundle):
                         rsc="R" score="INFINITY"
                     />
                 </constraints>
-            """
+            """,
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_primitive_avoids(self):
@@ -3247,7 +3260,8 @@ class BundleLocation(Bundle):
                         rsc="R" score="-INFINITY"
                     />
                 </constraints>
-            """
+            """,
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )
 
     def test_primitive_location(self):
@@ -3266,7 +3280,8 @@ class BundleLocation(Bundle):
                 <constraints>
                     <rsc_location id="id" node="node1" rsc="R" score="100" />
                 </constraints>
-            """
+            """,
+            output=LOCATION_NODE_VALIDATION_SKIP_WARNING,
         )